refactor(linkedin): extract endpoint cascade into processor registry
processLinkedInCapture had 5 if-branches over request.endpoint, each
40-60 lines with own dedup + store write logic. Deepen: each endpoint
becomes an EndpointProcessor behind a Record registry.

- linkedin/shared.ts: resolvePublicationId + SEARCH_ENDPOINT
- linkedin/process/{registry,feed,reactions,reposts,comments}.ts
- linkedin/index.ts: dispatch shrinks to 3 lines (harvest + lookup)
- search/process.ts: remove unused SEARCH_ENDPOINT import
- test/providers/linkedin/process-registry.test.ts

No behavior change. 169 tests pass, 0 fail, 4 snapshots identical.
