+++ src/providers/linkedin/index.ts
@@ -1,10 +1,3 @@
-import {
-  recordProvenance,
-  storeComment,
-  storeEngagement,
-  storePublication,
-  trackedHandleForProvider,
-} from "../../background/store";
 import type {
   BackgroundStore,
   ExportComment,
@@ -23,187 +16,14 @@ import type { CapturedPayloadMessage } from "../../shared/messages";
 import type { BackgroundProviderFacet, ScopeMode } from "../contract";
 import { publicationKey } from "../shared/utils";
 import { harvestSignature } from "./active-fetch/calibration";
-import {
-  linkedinFeedAccountInfo,
-  linkedinFeedToPosts,
-  linkedinFeedToPublications,
-  linkedinParseComments,
-  linkedinParseReactions,
-  linkedinParseReposts,
-} from "./parser";
-import { processLinkedInSearchCapture, SEARCH_ENDPOINT } from "./search/process";
+import { registry } from "./process/registry";
 import { searchScopeMode } from "./search/scope";
 
 export function processLinkedInCapture(store: BackgroundStore, request: CapturedPayloadMessage) {
-  // Harvest L3 (no SW — fonte única de verdade): colhe a assinatura volátil de toda captura
-  // LinkedIn. queryId vem da URL Voyager; clientVersion vem da signature (x-li-track). O csrf
-  // NÃO é colhido aqui — o scheduler o lê do cookie JSESSIONID em tempo de replay.
   if (request.url) harvestSignature(request.url, request.signature);
 
-  // A riqueza bespoke do LinkedIn vive em platforms.linkedin.extra (LinkedInExtra).
-  const lstore = store.platforms.linkedin.extra;
-  const handle = trackedHandleForProvider(store, "linkedin");
-
-  // Captura SDUI da busca (#14): roteada para o módulo dedicado em search/process.ts,
-  // que parseia o stream Flight, consolida no store e carimba a Provenance "search".
-  if (request.endpoint === SEARCH_ENDPOINT) {
-    processLinkedInSearchCapture(store, request);
-    return;
-  }
-
-  if (request.endpoint === "feedDashOrganizationalPageUpdates") {
-    const publications = linkedinFeedToPublications(request.payload, handle);
-    for (const pub of publications) {
-      // O filtro real (name OU headerText) vive no parser; aqui só gravamos a Provenance
-      // das publicações que já passaram por ele. Ver decisão 3 do plano #9.
-      if (handle) recordProvenance(store, "linkedin", pub.publication_id, "profile", handle);
-      storePublication(store, pub);
-    }
-
-    const posts = linkedinFeedToPosts(request.payload, handle);
-    for (const post of posts) {
-      lstore.posts[post.id] = post;
-      if (!lstore.feedOrder.includes(post.id)) lstore.feedOrder.push(post.id);
-    }
-
-    const accountInfo = linkedinFeedAccountInfo(request.payload, handle);
-    if (accountInfo) {
-      lstore.accountInfo = accountInfo;
-      store.trackedProfiles.linkedin = accountInfo;
-    }
-  }
-
-  if (request.endpoint === "socialDashReactions" && request.url) {
-    const result = linkedinParseReactions(request.payload, request.url);
-    if (result) {
-      if (result.isCommentReaction) {
-        const threadUrn = result.parentUrn;
-        if (!lstore.commentReactions[threadUrn]) {
-          lstore.commentReactions[threadUrn] = { users: [] };
-        }
-        const existing = new Set(
-          lstore.commentReactions[threadUrn].users.map((u) => u.provider_user_id),
-        );
-        const fresh = result.users.filter((u) => !existing.has(u.provider_user_id));
-        lstore.commentReactions[threadUrn].users.push(...fresh);
-      } else {
-        const parentUrn = result.parentUrn;
-        if (!lstore.reactions[parentUrn]) {
-          const total = 0;
-          lstore.reactions[parentUrn] = { users: [], total, lastStart: 0 };
-        }
-        const entry = lstore.reactions[parentUrn];
-        if (entry) {
-          const existing = new Set(entry.users.map((u) => u.provider_user_id));
-          const fresh = result.users.filter((u) => !existing.has(u.provider_user_id));
-          entry.users.push(...fresh);
-        }
-
-        for (const user of result.users) {
-          const pub = findLinkedInPublicationByUrn(store, parentUrn);
-          const pid = pub?.publication_id || parentUrn;
-          storeEngagement(store, {
-            provider: "linkedin",
-            publication_id: pid,
-            kind: "like",
-            engagement_id: publicationKey(
-              "linkedin",
-              `${pid}:like:${user.provider_user_id || user.username}`,
-            ),
-            actor: user,
-          });
-        }
-      }
-    }
-  }
-
-  if (request.endpoint === "feedDashReshareFeed" && request.url) {
-    const result = linkedinParseReposts(request.payload, request.url);
-    if (result) {
-      const parentUrn = result.parentUrn;
-      if (!lstore.reposts[parentUrn]) {
-        lstore.reposts[parentUrn] = { users: [], total: 0, lastStart: 0 };
-      }
-      const entry = lstore.reposts[parentUrn];
-
-      const existingUrns = new Set(entry.users.map((u) => u.urn).filter(Boolean));
-      const existingActivityUrns = new Set(entry.users.map((u) => u.activity_urn).filter(Boolean));
-
-      for (const user of result.users) {
-        if (user.urn) {
-          if (existingUrns.has(user.urn)) continue;
-          existingUrns.add(user.urn);
-          entry.users.push(user);
-
-          const pub = findLinkedInPublicationByUrn(store, parentUrn);
-          const pid = pub?.publication_id || parentUrn;
-          storeEngagement(store, {
-            provider: "linkedin",
-            publication_id: pid,
-            kind: "like",
-            engagement_id: publicationKey("linkedin", `${pid}:repost:${user.urn}`),
-            actor: {
-              provider: "linkedin",
-              provider_user_id: user.urn,
-              username: (user.name || "").toLowerCase().replace(/\s+/g, "_"),
-              name: user.name || "",
-              avatar_url: user.avatar_url || "",
-            },
-          });
-        } else if (user.activity_urn) {
-          if (existingActivityUrns.has(user.activity_urn)) continue;
-          existingActivityUrns.add(user.activity_urn);
-
-          const existingPost = Object.values(lstore.posts).find(
-            (p) => p.activity_urn === user.activity_urn,
-          );
-          if (existingPost) {
-            entry.users.push({ reshare_ref: existingPost.id, activity_urn: user.activity_urn });
-          } else {
-            entry.users.push(user);
-          }
-        }
-      }
-    }
-  }
-
-  if (request.endpoint === "socialDashComments" && request.url) {
-    const result = linkedinParseComments(request.payload, request.url);
-    if (result) {
-      const parentUrn = result.parentUrn;
-      if (!lstore.comments[parentUrn]) {
-        lstore.comments[parentUrn] = { items: [], total: 0, lastStart: 0 };
-      }
-      const entry = lstore.comments[parentUrn];
-      if (entry) {
-        const existing = new Set(entry.items.map((c) => c.comment_id));
-        const fresh = result.comments.filter((c) => !existing.has(c.comment_id));
-        entry.items.push(...fresh);
-      }
-
-      for (const comment of result.comments) {
-        const pub = findLinkedInPublicationByUrn(store, parentUrn);
-        const pid = pub?.publication_id || parentUrn;
-        comment.publication_id = pid;
-        storeComment(store, comment);
-        storeEngagement(store, {
-          provider: "linkedin",
-          publication_id: pid,
-          kind: "comment",
-          engagement_id: publicationKey("linkedin", `${pid}:comment:${comment.comment_id}`),
-          actor: comment.author,
-          engaged_at: comment.created_at,
-        });
-      }
-    }
-  }
-}
-
-function findLinkedInPublicationByUrn(store: BackgroundStore, urn: string) {
-  return Object.values(store.platforms.linkedin.publications).find(
-    (p) =>
-      p.provider === "linkedin" && (p.publication_id === urn || p.reposted_publication_id === urn),
-  );
+  const processor = registry[request.endpoint];
+  if (processor) processor.process(store, request);
 }
 
 function buildLinkedInCommentWithReactions(
